umq:change qbuf id to uint64_t

+++ src/urpc/umq/qbuf/umq_shm_qbuf_pool.c
@@ -410,7 +410,7 @@ static void umq_shm_qbuf_alloc_data_with_split(local_block_pool_t *local_pool, u
     bool first_fragment = true;
 
     QBUF_LIST_FOR_EACH(cur_node, &local_pool->head_with_data) {
-        uint32_t id = buf_to_id_with_data_split(pool->header_buffer, (char *)cur_node);
+        uint64_t id = buf_to_id_with_data_split(pool->header_buffer, (char *)cur_node);
         cur_node->buf_data = pool->data_buffer + id * umq_buf_size_small() + headroom_size_temp;
         cur_node->buf_size = umq_buf_size_small() + (uint32_t)sizeof(umq_buf_t);
         cur_node->headroom_size = headroom_size_temp;
